LPS-179383 Extract test class
